Fix issue with random images on billboards and weather search.
Random images now work. The issue before was that the random numbers list was not populated before it tried to find a random image, so it will use null as the link.
Issue with weather search was if city name had a space. Space was used to distinguish between city name and country, but if city had a space an issue was created. Found this issue with abu dhabi, so now country is found if there is a comma.